Include modified base service components
In order to experiment with the new rest url routing technique, I
included all modified base service components so that they can be
extended before merging these changes back.

This change moves the service path into the service factory, so that it
can later be used when extracting parameters from the url once the
service is invoked.